Added authentication

+++ project1/todos/todos.py
@@ -1,14 +1,14 @@
-from typing import Annotated
 from pydantic import BaseModel, Field
-from fastapi import APIRouter, Depends, HTTPException, Path
+from fastapi import APIRouter, HTTPException, Path
 from starlette import status
-from .database import engine, Session as SessionMaker
+from database import engine, db_dependency, auth_dependency
 from . import models
 from .models import Todos
-from sqlalchemy.orm import Session
+
 router = APIRouter()
 
 models.Base.metadata.create_all(bind=engine)
+
 
 class TodoReq(BaseModel):
     title: str = Field(min_length=3)
@@ -17,22 +17,8 @@ class TodoReq(BaseModel):
     priority: int = Field(default=0)
 
 
-def get_db():
-    db = SessionMaker()
-    try:
-        print("Yielding db session")
-        yield db
-    finally:
-        print("Closing db session")
-        db.close()
-
-
-db_dependency = Annotated[Session, Depends(get_db)]
-
-
 @router.get("/", status_code=status.HTTP_200_OK)
 async def get_all_record(db: db_dependency):
-    print("started quering the db!")
     return db.query(Todos).limit(20).all()
 
 
@@ -45,8 +31,10 @@ async def get_record(db: db_dependency, todo_id: int = Path(gt=0)):
 
 
 @router.post("/todo/create", status_code=status.HTTP_201_CREATED)
-async def create_record(db: db_dependency, new_todo: TodoReq):
-    item = Todos(**new_todo.model_dump())
+async def create_record(db: db_dependency, new_todo: TodoReq, user: auth_dependency):
+    if not user:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Not authenticated")
+    item = Todos(**new_todo.model_dump(), owner=user.get('id'))
     db.add(item)
     db.commit()
     db.refresh(item)
@@ -54,7 +42,9 @@ async def create_record(db: db_dependency, new_todo: TodoReq):
 
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_202_ACCEPTED)
-async def update_record(db: db_dependency, new_todo: TodoReq, todo_id: int = Path(gt=0)):
+async def update_record(user: auth_dependency, db: db_dependency, new_todo: TodoReq, todo_id: int = Path(gt=0), ):
+    if not user:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Not authenticated")
     item = db.query(Todos).filter(Todos.id == todo_id).first()
     if item:
         item.title = new_todo.title
@@ -69,11 +59,13 @@ async def update_record(db: db_dependency, new_todo: TodoReq, todo_id: int = Pat
 
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_202_ACCEPTED)
-async def delete_record(db: db_dependency, todo_id: int = Path(gt=0)):
+async def delete_record(user: auth_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if not user:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Not authenticated")
     item = db.query(Todos).filter(Todos.id == todo_id).first()
     if item:
         db.delete(item)
         db.commit()
-        return {"info":"Successfully deleted"}
+        return {"info": "Successfully deleted"}
     else:
         raise HTTPException(status_code=404, detail="Todo not found")
